requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.0
av==11.0.0
bidict==0.23.1
blender-qt-stylesheet==0.0.3
blinker==1.8.2
bqt==1.4.2
certifi==2024.8.30
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
contourpy==1.3.0
cors==1.0.1
ctranslate2==4.3.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
distlib==0.3.8
docopt==0.6.2
einops==0.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
faster-whisper==1.0.1
filelock==3.12.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.9.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
huggingface-hub==0.17.2
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.8
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
julius==0.2.7
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
lightning==2.3.3
lightning-utilities==0.11.6
llvmlite==0.40.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
numba==0.57.1
numpy==1.24.4
omegaconf==2.3.0
onnxruntime==1.19.0
openpyxl==3.1.3
optuna==3.6.1
packaging==23.2
pandas==2.2.0
parso==0.8.4
pillow==10.4.0
pipenv==2024.0.1
platformdirs==3.10.0
pooch==1.7.0
primePy==1.3
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
pure-eval==0.2.2
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==16.1.0
pycparser==2.21
Pygments==2.18.0
pyparsing==3.1.4
pyreadline3==3.4.1
PySide6==6.6.2
PySide6_Addons==6.6.2
PySide6_Essentials==6.6.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytorch-lightning==2.3.3
pytorch-metric-learning==2.6.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
QtPy==2.4.1
regex==2023.8.8
reportlab==4.2.2
requests==2.32.3
requests-file==2.1.0
resampy==0.4.2
rich==13.8.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
shiboken6==6.6.2
simple-websocket==1.0.0
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.6
speechbrain==1.0.0
SQLAlchemy==2.0.32
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tbb==2021.11.0
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tldextract==5.1.2
tokenizers==0.13.3
torch==2.4.0+cu121
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.4.0+cu121
torchmetrics==1.4.1
torchvision==0.19.0+cu121
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.33.2
typer==0.12.5
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
websocket-client==1.6.4
Werkzeug==3.0.4
whisperx==3.1.5
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
